+++ thorin/CMakeLists.txt
@@ -102,7 +102,8 @@ target_include_directories(libthorin PUBLIC
     ${CMAKE_CURRENT_SOURCE_DIR}/..
     ${PROJECT_SOURCE_DIR}/modules/half/include
 )
-target_link_libraries(libthorin PUBLIC absl::flat_hash_map absl::flat_hash_set)
+target_link_libraries(libthorin PUBLIC absl::flat_hash_map absl::flat_hash_set
+                                       absl::node_hash_map absl::node_hash_set)
 #if(MSVC)
     #target_compile_definitions(libthorin PUBLIC ABSL_CONSUME_DLL)
 #endif()
